♻️(persons) allow associating persons to organizations
Last week, we displayed the link between a person and its organizations
via their presence on the same course. The problem is that a course is
linked to several organizations and several persons so we are not able
to differentiate which persons belong to which organizations. We have
to add an explicit link between a person and the organizations to which
s.he belongs.