💄(search) prevent jumping from the "Clear filters" button
The "Clear x active filters" button in the search pane appears and
disappears depending on the number of currently active filters (0 or
any number).

Doing this in the render is fine but it causes the whole search pane
UI to jump around when the user selects the first filter or removes
it. We can fix this by replacing the render logic with a class and
`visibility: hidden`. This should also cause assistive technology to
ignore it.