✨(search) show filter values with a count of 0 as disabled
We started showing filter values with a count of 0 instead of plain
not having them appear at all, to prevent the search filters pane UI
from jumping around too much.

We don't want users to activate them as they would either do nothing
at all (if adding a value to a filter that already has active values)
or cause the search to have 0 results. We therefore need to disable
the checkbox and style its label to make its status obvious to users.