♻️(frontend) add utils to handle MPTT paths
We need to use MPTT paths to get information from them, such as
relationships between entities as defined by their CMS pages. This
is not a complex undertaking but we did make incomplete/incorrect
assumptions as we handled is casually at call sites.

Make a dedicated, unit-tested utils module to make sure we get the
correct information from MPTT paths.