🐛(search) fix an issue with non-drilldown filters behaving as drilldown
We encountered an issue whereby the languages filter, which is supposed
to be a non-drilldown filter (and is correctly marked `is_drilldown: false`
on the API), was behaving as a drilldown filter.

We narrowed this down to incorrect return values from our MPTT path helpers
that automatically remove parent/children filters of newly active filters
to simplify application state for the user.

Languages do not use MPTT paths as filter value keys, but were always
recognized as parent/children of each other. Fixing the MPTT path utils
to check for MPTT-path-likeness first solves the issue for any filter
that would use other kinds of keys, like language does (with ISO 639-1).