🐛(search) show all children for parent facets, not just top 5
We were only showing the top 5 children facets by matching course
count when a parent was supposed to show its children. This was due
to a faulty regex in the frontend that should have matched all the
parent's children but only matched some as it only covered numbers
in MPTT paths when it should have accounted for A-F (as they're in
hex).