✨(core) add admin utility to display readonly foreign keys as links
In Django admin, the readonly foreign keys are displayed as simple text.
This is broken as there is no easy way to access the related object.
This commit introduces a generic utility function that makes it easy
to display foreign keys in admin as links with custom anchor text, blank
target or not and even customer url computation.