✨(permissions) add an admin page for page roles
We need a way to see and create page roles from the admin.
We make the object readonly once created because page roles
only create permissions at the moment they are created. Once
created, the page role object is only usefull because it makes
it easy to find a page's group and page permissions.