✨(core) add factories to generate CMS pages
Up to now we have escaped doing it at the cost of convoluted code.
Creating page role instances will be much easier if we have a
factory to create pages. In future commits, we will have to refactor
the demo site and all other tests to use the page factory in place
of Divio's `create_page` utility function.