♻️(autocomplete) improve suggestion typings and simplify code
We have a clearer idea of what we need to achieve with Suggestions and
the kind of flexibility we need around them, now that we need to use
suggestions in different contexts.

Separate Suggestion types from their specific use in <SearchSuggestField />
in course search. Make the typings cleaner and more generic, and simplify
them to get rid of `suggestionsFromSection`.

In addition to improving the existing code, this makes it easier for us
to build other suggestion boxes.