🐛(search) fix course indexer when no related organization
If a course with no organization is indexed in Elasticsearch (which
should not happen if moderation is correctly done), the serializer
breaks because it expects at least one organization to be declared
on the course. Even if this is not a desired situation, the code
should not break so I fixed it and added a test to secure it.