♻️(wizards) use the DjangoCMS mixin for our wizard forms
In a future commit, we will need to know who is the user interacting
with a wizard. This is supported by DjangoCMS via a form mixin that
we overlooked. This commit makes sure all our wizard forms inherit
from this mixin and fixes all the tests that broke as a result.