✨(search) add persons index and search endpoints
We want to enable users to filter courses by instructors, to
make persons available in the general autocomplete, and, probably,
later on, to let users search for persons directly.

To achieve these, we need to index persons in their own ElasticSearch
index (through a PersonsIndexer and related utilities) and make a
ViewSet to allow the client to use that index for Search.