🐛(persons) fix person placeholders showing on the organization page
The person plugin was using the `page_placeholder` template tag with
the side effect that the organization page (which includes the person
plugin glimpse template) was showing the person's placeholders in its
list of placeholders in the plugin right side bar. Using the standard
`show_placeholder` template tag fixes it.