✨(search) index related person names on courses
We want users to be able to find related courses when they type
a related person's name in the course search main input, just
like we already do with categories and organizations.

To achieve this, we have to index these related persons' names
on the course and then search in them when the query comes in.