♻️(frontend) allow resource names (or kinds of objects) to be dynamic
We relied on hardcoded object kind names in the `modelName` enum. This
was useful to provide stricter typechecks on functions and object
indexing through the enum. The issue is that it now prevents us from
having dynamic object kinds, which we need to separately handle the
different kinds that can fall under "categories".

These kinds, such as eg. "Subjects" or "Levels" are dynamically
defined through the hierarchy of CMS pages on the backend side.
We therefore cannot rely on a hardcoded enum type to use them.