✨(search) add object kind field to autocomplete suggestions from API
We used to rely on MPTT paths used as IDs to determine which filter
a given suggested value is associated with. While this is "smart",
it prevents us from using suggestion kinds whose IDs are not MPTT
paths but regular IDs or UUIDs.

We simply added the object kind as a field to all suggestions to
solve this cleanly for "persons" and any future new kind of object.