✨(permissions) add permissions on Django Filer
We want users to be able to upload images but not anywhere. We decided
to create a folder in Django Filer for each organization and for each
course. Each folder is associated to its page and to the corresponding
user group via the PageRole object (so we can get a hold of it for
configuration and maintenance).