✨(search) add persons filter to courses Search
As we indexed Persons and added them as keywords on Courses, our
purpose was to allow users to find and filter courses by their
related persons.

Now that all the necessary groundwork has been laid, it's simply a
matter of adding a FilterDefinition and updating our defaults to
enable the persons filter.