✨(plugins) add a simple picture plugin for constrained templates
The way we use DjangoCMS on our specific page extensions does
not play well with the full DjangoCMS picture plugin. Indeed, all
image attributes should be controlled by code in the templates or
in settings and we should not bother users with all this.

The simple picture plugin only presents one field to the user to
upload a file. The way the picture is displayed in each location
is determined via a mega setting. Rendering options can be targeted
by the placeholder name or by a customer preset name passed via the
context.

Image presets are fully responsive with support for srcset.