♻️(courses) replace DjangoCMS picture plugin by simple picture plugin
As explained in the previous commit, the DjangoCMS picture plugin was
not adapted to the way we intend to collaborate on our specific page
extensions. Another problem was that up to now, all images uploaded
were displayed without any thumbnailing.

Our simple picture plugin kills two birds with one stone:
- low priviledge users can easily and safely upload an image at a
  placeholder that accepts it.
- images can be resized and cropped specifically for each place where
  they are used... and this is fully controlled by code.