✨(models) categories can retrieve their own meta category
As we use categories, we might find ourselves needing to find the
meta category under which a given category is filed. In most cases,
it's most practical if we can do so without referring to any
external information.

We simply add a method on Category that issues a request that finds
the only public category that has no category ancestor, among the
current category's ancestors. This is by definition its meta category.