🔧(courses) add default settings for the simplepicture plugin
This is different from what is done in the defaults.py file which
defines default value for settings introduced by the courses app.
The settings added by this commit add default values for settings
introduced by other or third party apps because the courses app
requires a minimum configuration to work.

The pattern we propose here allow to use these defaults either with
the usual Django pattern: `from settings import *`, or with Django
Configuration via a mixin that we provide.