✨(courses) add template tags to keep img declarations in templates
The simple picture plugin allowed to fully customize and optimize with
code  how our images are displayed. One issue raised during review was
that everyting is defined in settings which is not convenient. It
makes more sense to fine tune img attributes in templates.

The first template tag allows to retrieve a placeholder's plugins
instead of their rendered HTML, in order to render them inside the
template with the second block template tag.