✨(search) add the "kind" field to categories during indexing
We need to build autocomplete for categories based on filtering
them by kind. This means we need a "kind" field on the categories
to enable that filtering on our completion or search requests.

We can use the reverse_id of the related meta category page to
tie together our filters, endpoints and query strings through
a neat machine name.