✨(search) enable autocomplete dynamically based on settings
Autocompletion for filters in courses search was hardcoded through
strings in an array in the <SearchSuggestField /> component. This
is not appropriate as we want to make richie easy to personalize, and
it also breaks down if a user disables the filter related to one of
the hardcoded autocompletion filters.

We can instead add an `is_autocompletable` flag on filter definitions
to make them truly dynamic. This allows the frontend to only offer
autocompleted suggestions for filters which provide API endpoints
and are enabled in the backend.