♻️(courses) move richie specific settings to the courses app
A lot of the settings in the sandbox are sensible defaults specific
to Richie. We want to make it easy to start a Richie project by
importing default settings so we need to distribute them with the
app.

The pattern proposed here allows to import our settings with import *
as it is done on most Django projects... but it is also possible to
import them in Jazzband's Django Configuration via a mixin.