🐛(permissions) fix page permissions that were making pages private
The PagePermission model is actually a 2-headed snake... it can act
as a permission if `can_view` is False but acts as a view restriction
if `can_view` is True. Said differently, the purpose of the `can_view`
flag is not to precise that staff users have the right to view the
page... its purpose is to define who can view the page on the
frontend.

What we want here is a permission on admin pages for our staff users
so `can_view` should be False.