♿️(search) improve label for child filter toggle
The child filter toggle in the course search was deemed compliant,
but the label for the toggle is not very explicit, especially for
a screen reader user that cannot see its context.

Instead of "Show child filters", we can have it say "Show more filters
for Subject X", which is way more explcit. Ditto for the other toggle
position.