✨(plugins) extend the simple text plugin to allow more configuration
We want to be able to configure for a given placeholder, what formatting
options are available on CKEditor and a max length. The way we implement
it here is in the spirit of DjangoCMS placeholder configuration. If the
current placeholder slot name matches a configuration in our setting,
we use it.