📝(README) update README to the latest situation
We forgot to remove the reference to Open edX in the title. The
project is now LMS agnostic. I also remove the claim of "multi-site"
because, even though DjangoCMS is natively multi-site, the Richie
search engine would need more work to really support multi-site
indexes. It is not something we plan to do as we prefer to run many
separate Docker micro services.