♿️(frontend) improve course glimpse links
Our course glimpse links have two accessibility-related issues. They
have useless and non-conforming titles, and they contain images with
non-empty alternative text.

This means there is a lot of redundancy in the markup, and causes
2 non-conformities for WCAG/RGAA.

We can simply force empty alt text for the image in the course glimpse
and remove the link title entirely.