♻️(templates) improve course glimpse fragment use
As we were modifying links in course-glimpses to make them compliant
with WCAG/RGAA, we noticed there was a structural problem in the markup.
We were putting empty alts on images in the course-glimpse fragment
while the links that justified this empty alt were located in the
fragment call sites in other templates.

This made us realize the course glimpse link actually belongs in
the course glimpse fragment along with the rest of the course glimpse
markup.

We also rationalized some CSS and removed 🔥 some useless classes
along the way.