🐛(search) don't keep stale user input when a suggestion is selected
As the user inputs text into the autocomplete search bar, it is used
for our search-as-you-type interface, and also send to the API to
compute suggestions.

When a user selects a suggestion, we need to remove the query added
to params by search-as-you-type because the text entered by the user
was already "consumed" by the action to select the suggestion. It is
stale and the UI does not reflect it anyway.