♿️(blogposts) fix non-conformities in blogpost glimpses
Like our other kinds of glimpses, blogpost glimpses contained
non-conformities related to links and images.

As we worked to fix them, we noticed an issue with the template
boundaries, as part of the blogpost glimpse interface elements
(the <a> tags) lived in the call sites instead of the blogpost
glimpse template itself.

We updated these structures to move the <a> tags to their more
logical place.