✨(search) increase the number of counted facets for specific searches
By default, ElasticSearch returns the 10 top terms when making
an aggregations call on top terms. This fit us pretty well when we
just show the top values by default.

However, it's not enough when we run a search for a specific number
of facet counts: we want the client to be able to request all the
facets they need and get them (within reasonable bounds).

We added settings to manage these values (default and max facet limits),
and wrote some tests for this behavior.