♻️(roles) factorize create page role for an organization
We were creating the page role for an organization only when it was
created via a wizard. We need to also create the page role in the
demo site or when creating organizations programmatically. This is
made easy by factorizing the code that creates the page role and
associated group, permissions and filer folder.