♻️(roles) factorize create page role and permissions for a course
We were creating the page role for a course and permissions for its
related organization only when it was created via a wizard. We need
to also create do it for the demo site or when creating courses
programmatically. This is made easy by factorizing the code that
creates the page role, associated group, permissions and filer
foler.