💄(courses) add styles for category icon & color on course glimses
The icon was added without styles on course glimpses. We needed
to add some styles so that it was displayed in an acceptable way
that does not break the app flow.

We also added a purely decorative element that uses the category
color to display a colored band on course glimpses.