💄(search) move the search template title into react
Currently the Django template is responsible for displaying the page
title on the search results, like on any other page. This is good in
that it allows this title to be managed in the CMS. However, it is
troublesome in terms of page structure as we want to handle the
search filters pane more cleanly.

To keep the title in the CMS while improving our page structure, we
decided to pass the page title to the React frontend via a data
attribute in the Django template. This lets us cleanly insert it in
the relevant place in our markup.

In our opinion, this is a better compromise than hardcoding it in
React (as we lose CMS management of the title, which should not
happen if we can avoid it), or adding hacks to the CSS which can
be more harmful in the longer term.