✨(search) show a modal to search for more possible filter values
In the search filters pane, we display the top N (by default 10) values
for each filter, ordered by facet count. It makes sense to show the
values with the most matching courses more prominently, but users
might want to filter by other values, especially on Richie instances
that host a large number of courses.

This is where <SearchFilterGroupModal /> comes into play. It's shown
as a button below the top facets that opens a modal that lets users
search for more values for the related filter and add them to their
current course search.