💩(search) wrap Richie's <body> in a div to hide when a modal is open
It is important for users of screenreaders that other page content be
hidden (via the aria-hidden attribute) while the modal is open. To
allow react-modal to do this for us, we need the page content to be
wrapped in an easy-to-select element that does not also contain the
modal itself, hence our #modal-exclude addition.

As we handled this requirement, we ran into issues with our testing
environment, which we only managed to overcome through a dirty
workaround for now. The issue is that the `setAppElement()` call
expects a real element to exist, which is not the case in test
environments.