♿️(css) add a css class to move content offscreen for accessibility
In some cases, like our upcoming Pagination for the course search,
we need to display some textual content only for users of accessibility
helpers. This is content that provides context that other users can
infer visually.

For those use cases, the best way to handle this content is to simply
position it offscreen. Screen readers or other such assistive tools
will not ignore them, but they won't visually appear on the screen.