✨(frontend) update course search params hook to manage page changes
As we add pagination to our course search view, we need to manage
the related params, which are handled by our useCourseSearchParams
hook.

We can simply add an action to our reducer that handle page changes
by updating the offset in the query parameters.

We also needed to update existing actions that affect the content of
the query to also reset the pagination by forcing offset to 0. This
prevents users from being shown non-optimal or non-existent pages
of results if they change their query after having browsed through
pagination.