✨(search) scroll to the top when the user moves pages
When the user moves through pages in our course search engine, they
would expect to see the top of the new page, now the bottom of it. It
will also make it clearer that they did in fact change pages. We can
achieve this with minimal effort through the use of window.scroll.

Note: we don't want the same behavior to apply when user filters, even
if technically we're moving pages, because we have no way of knowing
if the user is done interacting with filters and it would be a bad
experience to be forcibly scrolled to the top in those instances.