✨(search) paginate the course search view
On large richie instances, there can be hundreds of courses available
to display. As the default view of the search page is non-filtered,
the page could be slow to load with hundreds of courses to load from
the API and display in complex glimpses with their cover images.

Instead, we can paginate the view down to a useful level (say,
20 courses by default) which keeps the page reasonably long and
deals with performance issues.