♻️(frontend) update our code for react-intl v3+
Versions 3+ of react-intl rely on browser native implementations of
Intl for major parts of the localization workflow. This means, for
example, that compatible browsers no longer need to download
locale-sopecific data to format dates, pluralize, etc.

Still, we need to support less-modern browsers that do not include
these features yet, which is why we're adding polyfills in our index,
that are only loaded when the features react-intl needs are missing.