🐛(search) make child filter values unfolding dynamic
Search filter values that have active children automatically
unfolded on initialization, then only relied on user input
to fold/unfold after that.

This is not ideal when we're adding children filters from the
autocomplete search bar or one of the smaller search fields as
the user is not shown the filter value they just added.

We made it dynamic (by checking if there is an active child on
each re-render instead of only at initialization) and gave priority
to the uder choice over the automatic unfolding.