✨(search) make autosuggestion diacritics insensitive
Autosuggestion API endpoints treat accented characters as different
from the non-accented versions of the same characters. This is
technically correct but is not intuitive for users who want to
see the results they expect regardless.

We can create a custom analyzer that folds accented characters into
their ascii version and use that for all our completion fields in
our object mappings.