✨(courses) allow icons to display on course detail and not on glimpse
We added a separate placeholder to attach category icons to a course. With
this design, it was not possible to add an icon to a course without it
appearing on the glimpse. In order to allow this use case, we refactored
categories to display as icons when they have an image, regardless of the
placeholder to which they are added on the course page.

The first category in the "icon" placeholder is still displayed on the glimpse
but t is now possible to add icons to the course detail view via the old "tag"
placeholder so that they don't appear on the glimpse.