♻️(search) merge the TermsAggsMixin into the IndexableFilterDefinition
The `TermsAggsMixin` was originally implemented as a mixin to make it
easier to understand and to reuse, even though its only call site was
in `IndexableFilterDefinition`.

Now however, we need to make changes to the `get_aggs_fragment` method
in the `TermsAggsMixin` that interact with changes to the
`get_faceted_definitions` method in IndexableFilterDefinition. It
therefore makes no sense to keep these methods on separate classes
when they are in fact coupled.

We can just move `get_aggs_fragment` from TermsAggsMixin onto
`IndexableFilterDefinition` and delete the now empty `TermsAggsMixin`.