✨(search) add a "has_more_values" flag on filters in course search
We added a way to search for more values for filters in the filter
panel through the use of a button and modal. This button and modal
should notably not be shown when there are no more available values
than what the panel is already showing.

To achieve this, we need to add this information on response to API
calls on course search (which is where filter definition building
occurs). This is not straightforward to determine for all filters,
which is why we had to test all cases and add some slightly
convoluted code to add this flag to faceted filter definitions.