✨(search) add information on searchability in filter definitions
Filter definitions already include information on whether or not
their values can be autocompleted, which it uses in its completion
interfaces.

To properly assess whether it needs to show the "View more" button
and modal on filters, the frontend not only needs to know if there
are more values than what it already has, it also needs to know if
a search endpoint exists for this filter.

This is exactly what the `is_searchable` flag should achieve.