✨(search) only show the "More options" button on filters when necessary
There is no point in showing the "More options" button or opening the
related modal for a filter group when the related filter is not
searchable of all possible values are already displayed.

We simply make use of the new flags on the objects in the API to
only show the button and modal when they are actually useful.