🐛(courses) fix a template error in the person plugin
The person plugin, when used through its tag form, contained an
error where `get_title` was called instead `get_title`. It was also
called on the person object (which has no such method) instead of the
related page (which does, and has the person's name as its title).

We simply fixed the error and added tests to make sure it works from
now on.