♻️(search) fix "more options" modal for non-mptt-paths based filters
The "more options" modal for filter values in the search filters pane
incorrectly made the assumption that all searchable filters were based
on MPTT paths and built a regexp accordingly.

That is not the case: in our own sandbox environment, persons are not
using those paths and the "more options" for persons modal is broken.
We can simplify the regexp to make it work with every filter.