♻️(search) improve error messages in the "more options" modal in filters
The "More options" modal for filter values was making request to the
server as soon as the user started to type, and immediately showed a
cryptic error message as it received an error for the server, because
textual search only works in our API endpoints with 3 characters or more.

Stop sending useless API requests when users input 1 or 2 characters and
show a helpful message in the modal instead.