♻️(search) improve and complete strings in pagination
Visible pagination does not contain a lot of strings, but we have
a few more strings for accessibility helpers that provide context
along with the page numbers.

We previously attempted to do this "smartly" through the HTML
structure of our interpolated strings. This is not ideal as it works
for english and some other languages, but fails for other languages
with a different grammar.

Also, we forgot a bunch of strings, which resulted in hardcoded
english in our template.