♻️(front) refactor <SearchSuggestField /> to make parts reusable
The <SearchSuggestField /> component is responsible for the textual search
and autocompletion field Richie offerts on its Search page, which is its
only React-rendered part for now.

Going forward, we want to display a similar autocompletion field in other
parts of the website, without the surrounding search results and filtering
capabilities.

Since the UI is similar, we can share some logic. However, the glue that
plugs it with our search engine is not something we want to reuse for the
universal field.

Therefore, we're pulling all the reusable logic out of <SearchSuggestField />
and into a shared helpers file. This way, we can cleanly reuse those parts in
our incoming universal field.