✨(program) rename 'description' to 'excerpt' and add 'body' to contents
Previously there was only a single text content named 'description'. But it has
potential to break glimpses. We need distinct text field for glimpse and detail.

So we renamed 'description' to excerpt which still a PlainTextPlugin to use in
glimpses then added a 'body' to accept a TextPlugin which is used in details.