✨(layout) add breadcrumbs on top of page details
Pages missed breadcrumbs which is a penalty for ergonomy, since we have more
than two page levels.

Here we add breadcrumbs as a common element from base template skeleton in
its own template block so on default it is enabled everywhere. However we do
not want to show it on homepage and first level page which are already
visually active in header menu, so we disable breadcrumbs block from all
page listing templates.

This does not involve any new tests since we stand on basical feature from
Django templates and DjangoCMS.