⬆️ (circle) upgrade crowdin image version
The circle-ci job uploading translations on crowdin is failing since few
days because there is an issue with the tar version used in our crowdin
image. We think it's a circle-ci regression, they remove the fallback on
their tools to use the good tar commad. We decided to add tar in our
crowdin image and to use it in our upload-i18n-strings job.