🐛(static) fix logo location that breaks static finder
The logo image location was modified but the SEO tag PR was opened
before this and still had the old path. This error only appears
after deployment because it is caused by the use of the Manifest
static storage backend. We should add tests in our CI to identify
this type of issues earlier.