👌(layout) fixed "srcset" attribute value on every "<img>" tags
It has been reported HTML validator errors about invalid "srcset" attribute
value on "<img>" tags. Size item list must not ends with a "comma" but due
to the list is builded from templates, it often lead to ending with a comma.

This commit update every templates with "srcset" attribute on image tags to
avoid ending on a comma in size item list.