👌(layout) add default djangocms_video template to fix invalid HTML
It has been reported HTML validator errors about invalid obsolete
"frameborder" attribute value on "<iframe>" tags.
It is only related to djangocms_plugin templates.

This commit adds a default template for the "djangocms_video" plugin to
fix its "<iframe>" tag since it is the only way to fix it for every plugin
usages.