🏷️(frontend) fix intl type breakage from dependency
We relied on having a type definition for `Intl.RelativeTimeFormat`
as we believed it existed in the typescript libs. It was really
coming from one of our dependencies extending `Intl` with it.

We can just type assert as the only call that uses `RelativeTimeFormat`
is to assert whether or not to run a polyfill.