♻️(frontend) add common data props expectation to root components
We started up data-passing interfaces between the backend and frontend
through a JSON property put on the root DOM element and parsed by our
own index.tsx before components are instantiated.

In addition to ad-hoc properties for any component, we're starting an
interface on all our "root" components (those that can be called directly
from the backend) to pass them general data about the app. The first
order of business being access to statics: the React frontend could not
access them in any way because of Manifest file storage.