🚸(search) make the search filters pane a sliding drawer on mobile
While it works as expected on larger screens, the search filters pane
occupies a lot of space on mobile, where users have to scroll through
it to get to the course search results. This is not ideal.

We decided to make it a sliding drawer on smaller screens. It's hidden
by default and slides in when the user clicks on a "filter" button.
It can then be closed again to interact with the search results.