🏗(hybrid) experiment with a way for the frontend to access assets
Our React powered frontend currently does not have any access to
Richie's assets, as Django's manifest file storage hides them
with a hash for cache busting purposes.

The only realistic solution is to pass the path to necessary assets
to the frontend. Here, we experimenting with doing it to the
data-props that all root frontend components are supposed to accept
and render with.