✅(persons) improve tests for the person detail view
We were not differentiating between a person that had never been published
and a person that had bee published and then unpublished. We were also not
checking that draft content was only shown to authorized users. I added tests
to secure the same behavior as on organizations but the way it works is not
completely coherent so I opened an issue for improvement:
https://github.com/openfun/richie/issues/843