♻️(category) include descendants in related objects
When displaying related objects for a category, we were only displaying
the objects actually linked to the current category. For all the
categories that are not leaves, this was resulting in displaying no
related objects at all. What we want is to display the objects related
to either the current category or one of its descendants.

After doing so, I realize that the subcategories of a category should be
displayed on top of the page, before all the related object sections.