🔥(category) remove anti-pattern in factories to create related objects
We were creating courses related to a category from its factory. This was
confusing because the "should_publish" option did not impact the course
object so the category was published but its relation with the course was
still draft. There is not easy way to fix it as we sometimes want to create
a draft relation, and other times a published relation. The best is to
only use fill_* methods directly on the object that carries the relation,
as demonstrated in the test we fix.