👌(layout) droped usage of "<button>" in course glimpse
It has been reported HTML validator errors about button element inside link
element which is forbidden because button and link elements are both expected
to be a "call to action" this alike putting a link inside a link.

Since the button element usage was not really useful and its layout was
already done just from "button" class, the button element has been changed to
a span element, keeping the "button" class.