♻️(frontend) load global polyfill with imports instead of an entrypoint
We were loading our global polyfills using entrypoints in our webpack
config. This worked well until now.

However, as we want to enable users to make their own build of Richie,
this is problematic, because paths to entrypoints are fixed. We decided
to use them through imports in our index so they can be resolved to
their actual path at build time, depending on the context.