♻️(frontend) move to absolute path imports
We used relative import paths until now even though absolute paths are
somewhat cleaner and simpler to understand. We did so because it was
what we adopted at the beginning of the project and because it is
simple config-wise.

Now, we decided to move to absolute imports to make it easier to override
components in dependent projects, and to make it easy for us to make sure
we only override components.