🔧(frontend) directly require babel conf in webpack conf
We used to rely on automatic loading of our babel config file by
babel when building the javascript frontend in Richie. This worked
as long as we were only building it directly.

Now that we need to support builds run from Richie dependents, it's
way easier for that use case if we explicitly load our babel config,
otherwise users would have to copy or symlink it into their projects.