🔧(frontend) update the webpack config to enable override by dependents
Add support for frontend component overrides through custom webpack
command line args. Dependent builds can now specify a settings file which
will be loaded and used in the build (though it only handles overrides for
now). To allow this to function we also added a flag to enable path
resolution to the dependent's own code and node_modules.