👷(cookiecutter) run linters on cookiecutter template
We want the Cookiecutter template to be valid Python code so we should
run our linters on it. I managed to add black and bandit (plus isort
and flake8 were already running on it) but I could not run pylint on
it because the Python code is not able to execute on the template.