👷(circleci) fix test-back-mysql following mysql image upgrade
A new mysql:5.6 image was published that broke our CI because our fun
user was not allowed to create a database anymore. It is necessary
because tests run on a specific database called `test_richie` that it
creates if it is missing. Making sure the database is already created
in the container when it starts fixes the bug.