📝(website) create a documentation & "marketing" website for Richie
Richie was documented through markdown files in a docs/ folder in the
repository, and publicized through the README.md. As we're starting to
grow the community, this is not enough.

We're starting a documentation website with a decicated "marketing"
landing page to introduce new users/contributors to the project. We
used Docusaurus to automatically build out the documentation using
our own markdown files for content.