👷(docs) add a CircleCI job to build the docs on master
We want to automatically keep the published documentation website
up-to-date with the repository. To achieve this, we can use a job that
runs only on master to build and publish the website to GitHub Pages.

It should be sufficient as any new release of Richie should include
a cut of the documentation website in the release commit merged on
master before tagging.