🐛(gh-pages) add a CNAME in gh-pages root to keep the deployment domain
Even though it appears as a setting in the Richie repo on GitHub, the
gh-pages custom domain is actually just a file at the root of the
repo in the `gh-pages` branch. The UI in the settings just changes this
file. This is then overwritten by our deployments.

To avoid losing the custom domain with every deployment, we can simply
add the CNAME file to our statics so it appears at the root in the
`gh-pages` branch after our deployments.