🐛(website) fix documentation build
Automated deployment to gh-pages in the CI fails because it runs with
HTTPS and therefore hangs on an HTTP password prompt instead of using
the git user token we generated for funmoocbot.

The `USE_SSH` environment variable should force it to use ssh to push
to github and get rid of this password prompt.