♻️(interop) rename fun-react conventional class to richie-react
When we started rendering React components from Django templates, we
needed a class to allow react to find which parts of the page it needs
to take over.

We used `fun-react` for this. As Richie is living on its own as an
open-source project, it seems more appropriate to us if this class
is named `richie-react` instead.