📝(interop) improve Django-React interop documentation
Documentation on the interoperability of React components and Django
templates dated from the first time we built this interop, and was
missing key features such as data-locale and data-props.

As we recently built a full documentation website, it seemed necessary
to us to improve on this by completing the documentation and adding a
list of built-in available React components for Richie.