🐛(frontend) fix a 400 Bad Request issue in <RootSearchSuggestField />
To initialize itself and know which filters can be used for
autocompletion, <RootSearchSuggestField /> waits for valid user
input and makes a request on the course search endpoint with a
`"filters"` scope.

This request was failing due to incorrect case on the `scope`, which
needs to be `"filters"` instead of `"FILTERS"`, and an incorrect
value for `limit`, which can't be 0 even though we're not expecting
any actual search results.